last updates for Order Routes

+++ controllers/ordersControlers.js
@@ -1,10 +1,12 @@
 
 let Orders = require("./../models/orderModel");
-const {orderUpdate } = require("../middlewares/orderUpdate");
+const { orderUpdate } = require("../middlewares/orderUpdate");
+// const { body, params, query } = require("express-validator");
 
 
 module.exports.getAllorders = function (request, response,next) {
-    Orders.find({})
+    Orders.find({}).populate({ path: "userId", select: "firstName lastName email mobile address" })
+        .populate({ path: "product", select:"name price"})
         .then((data)=> {
             response.status(200).json(data);
         })
@@ -18,9 +20,14 @@ module.exports.getAllorders = function (request, response,next) {
 module.exports.getOrderbyID = function (request, response,next) {
     Orders.findOne({ _id: request.params.id })
         .then(data => {
-            if (data == null) { next(new Error("This Order is not found..!")); } else {
-                response.status(200).json(data);
-            }
+           if ((request.role=="admin") ||((request.role == "user") && (request.id == data.userId))) {
+                if (data == null) { next(new Error("This Order is not found..!")); } else {
+                    response.status(200).json(data);
+                    console.log(request.id)
+                }
+            } else { next(new Error("Not authorized!")) }
+            
+            // console.log(params.id)
         })
         .catch(error => next(error));
    
@@ -30,12 +37,13 @@ module.exports.addOrders = function (request, response,next) {
         _id:request.body._id,
         userId: request.body.userId,
         product: request.body.product,
-        price: request.body.price,
+        totalPrice: request.body.totalPrice,
         quantity: request.body.quantity,
         Status:request.body.Status
     });
     object.save()
         .then(data => {
+            
             response.status(201).json({ data: "This Order is added succecfully..!" });
     }).catch(error=>next(error))
         
@@ -45,11 +53,31 @@ module.exports.addOrders = function (request, response,next) {
 module.exports.updateOrders = orderUpdate;
 
 
-module.exports.deleteOrders = function (request, response) {
-    Orders.deleteOne({ _id: request.params.id }
-        ).then(data => {
-                response.status(200).json(data);
-            })
-        .catch(error => next(error))
-   
+module.exports.deleteOrders = function (request, response,next) {
+    Orders.findOne({ _id: request.params.id })
+    .then(data => {
+       if ((request.role=="admin") ||((request.role == "user") && (request.id == data.userId))) {
+            if (data == null) { next(new Error("This Order is not found..!")); } else {
+                Orders.deleteOne({ _id: request.params.id }
+                    ).then(data => {
+                            response.status(200).json(data);
+                        })
+                    .catch(error => next(error))
+            }
+        } else { next(new Error("Not authorized!")) }
+        
+        // console.log(params.id)
+    }).catch(error => next(error));   
 };
+
+
+
+// module.exports.deleteOrders = function (request, response) {
+    
+//     Orders.deleteOne({ _id: request.params.id }
+//         ).then(data => {
+//                 response.status(200).json(data);
+//             })
+//         .catch(error => next(error))
+   
+// };
